requirements.txt
torch==2.11.0
torchvision==0.26.0
numpy==2.4.4
pillow==12.2.0
opencv-python==4.13.0.92
scikit-learn==1.8.0
tqdm==4.67.3
Flask==3.1.3